(closes #64) Extract class, support href-to on components.
The purpose of this PR was to enable the `href-to` functionality to work
on components:

```hbs
{{#my-button style='link' href=(href-to 'about')}}
  About
{{/my-button}}
```

Originally, href-to ignored all components by filtering out targets
with `.ember-view`. This was added so the `href-to` handler wouldn't
interfere with instances of LinkComponent. However, we want to support
the above use case.

After updating the `canHandle` code in the original initializer, it was
hard to test the new behavior (because of the dependency on the registry).
We also thought the initializer logic was becoming complicated.
So, we extracted the logic into a separate class, and added some more test
coverage along the way.